Split project into WASM modules
