+++ src/input/mod.rs
@@ -1,0 +1,3 @@
+pub mod camera;
+pub mod keyboard;
+pub mod mouse;
