+++ src/web.rs
@@ -1,0 +1,64 @@
+use std::{cell::RefCell, rc::Rc};
+use wasm_bindgen::prelude::*;
+use wasm_bindgen::{closure::Closure, JsCast};
+
+use glam::Mat4;
+
+use crate::render::state::State;
+use crate::input::camera::Camera;
+use crate::input::{keyboard, mouse};
+
+#[wasm_bindgen(start)]
+pub async fn start() -> Result<(), JsValue> {
+    console_error_panic_hook::set_once();
+    let window = web_sys::window().unwrap();
+    let document = window.document().unwrap();
+    let canvas = document
+        .get_element_by_id("gpu-canvas")
+        .unwrap()
+        .dyn_into::<web_sys::HtmlCanvasElement>()?;
+
+    let state = Rc::new(RefCell::new(State::new(&canvas).await?));
+    let performance = window.performance().unwrap();
+    let aspect = state.borrow().aspect;
+    let camera = Rc::new(RefCell::new(Camera::new(aspect)));
+
+    keyboard::attach(&window, camera.clone());
+    mouse::attach(&window, camera.clone());
+
+    let start_time = performance.now();
+    let prev_time = Rc::new(RefCell::new(start_time));
+    let f: Rc<RefCell<Option<Closure<dyn FnMut()>>>> = Rc::new(RefCell::new(None));
+    let g = f.clone();
+    let window_c = window.clone();
+    let perf_c = performance.clone();
+    let camera_c = camera.clone();
+    let state_c = state.clone();
+    let prev_time_c = prev_time.clone();
+
+    *g.borrow_mut() = Some(Closure::wrap(Box::new(move || {
+        let now = perf_c.now();
+        let dt = (now - *prev_time_c.borrow()) as f32 / 1000.0;
+        *prev_time_c.borrow_mut() = now;
+        let elapsed = (now - start_time) as f32 / 1000.0;
+        let angle = elapsed / 5.0 * (2.0 * std::f32::consts::PI);
+        {
+            let mut cam = camera_c.borrow_mut();
+            cam.update(dt);
+            let cam_matrix = cam.matrix();
+            let model = Mat4::from_rotation_z(angle);
+            let mvp = cam_matrix * model;
+            let mut st = state_c.borrow_mut();
+            st.update(mvp);
+            if st.render().is_err() {
+                return;
+            }
+        }
+        window_c
+            .request_animation_frame(f.borrow().as_ref().unwrap().as_ref().unchecked_ref())
+            .unwrap();
+    }) as Box<dyn FnMut()>));
+
+    window.request_animation_frame(g.borrow().as_ref().unwrap().as_ref().unchecked_ref())?;
+    Ok(())
+}
